Mark 3
I have made changes to the Preliminary model so it now includes a
weighted option that is chosen by the user. I also cleaned things up by
making a separate function to calculate the WMA (weighted moving
average). If input is not correct for weights, the function will default
to GMA.