Debugged Profit
I think the profit equations are debugged. I changed them to make sure a
'sell' only happens after a 'buy' and the function will not 'sell'
unless the 'buyAmt' is less than the 'sellAmt': THIS IS  A MAJOR
ASSUMPTIONS THAT ASSUMES THE STOCK WILL CONTINUE TO GROW AND DOES NOT
HAVE  A "CUT-YOUR-LOSES-FAIL-SAFE".  I am thinking about putting the
fail safe in but wanted to have a group discussion first.